+++ core/src/server/Server.cpp
@@ -193,9 +193,10 @@ Server::Start() {
 #else
         SERVER_LOG_INFO << "CPU edition";
 #endif
-        LogConfigFile(config_filename_);
-        LogHardware();
-        LogConfig();
+        /* record config and hardware information into log */
+        LogConfigInFile(config_filename_);
+        LogCpuInfo();
+        LogConfigInMem();
 
         server::Metrics::GetInstance().Init();
         server::SystemInfo::GetInstance().Init();